requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ecdsa==0.19.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
kombu==5.6.2
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.3.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.42.0
vine==5.1.0
wcwidth==0.6.0
